+++ src/index.js
@@ -1,1 +1,2 @@
+export * from './MorphingModal.svelte';
 export { default } from './MorphingModal.svelte';